+++ opengwasdb/index/__init__.py
@@ -4,20 +4,16 @@ from opengwasdb.index.analyses import AnalysesIndex
 from opengwasdb.index.sqlite import (
     connect,
     count_rows,
-    create_lookup_indexes,
     get_metadata,
     initialise_schema,
     set_metadata,
-    variant_by_identifier,
 )
 
 __all__ = [
     "AnalysesIndex",
     "connect",
     "count_rows",
-    "create_lookup_indexes",
     "get_metadata",
     "initialise_schema",
     "set_metadata",
-    "variant_by_identifier",
 ]